refactor(server): isolate aggregate pandas implem
